+++ ZAPD/ZRoom/Commands/SetObjectList.h
@@ -10,7 +10,7 @@ public:
 	virtual std::string GenerateSourceCodePass1(std::string roomName, int baseAddress);
 	virtual std::string GetCommandCName();
 	virtual RoomCommand GetRoomCommand();
-	virtual int32_t GetRawDataSize();
+	virtual size_t GetRawDataSize();
 	virtual std::string GenerateExterns();
 
 private:
