+++ ZAPD/ZVector.cpp
@@ -53,11 +53,11 @@ void ZVector::ParseRawData()
 	assert(this->scalars.size() == this->dimensions);
 }
 
-int ZVector::GetRawDataSize()
+size_t ZVector::GetRawDataSize()
 {
-	int size = 0;
+	size_t size = 0;
 
-	for (int i = 0; i < this->scalars.size(); i++)
+	for (size_t i = 0; i < this->scalars.size(); i++)
 		size += this->scalars[i]->GetRawDataSize();
 
 	return size;
@@ -93,7 +93,7 @@ std::string ZVector::GetSourceValue()
 {
 	std::vector<std::string> strings = std::vector<std::string>();
 
-	for (int i = 0; i < this->scalars.size(); i++)
+	for (size_t i = 0; i < this->scalars.size(); i++)
 		strings.push_back(scalars[i]->GetSourceValue());
 
 	return "{ " + StringHelper::Implode(strings, ", ") + " }";
